Use standard definition of rint and fix pointer error

+++ src/engine/qcommon/q_math.c
@@ -4185,7 +4185,7 @@ void QuatTransformVector( const quat_t q, const vec3_t in, vec3_t out )
 }
 
 #if defined(_WIN32) && !defined(__MINGW32__)
-int rint( double x )
+double rint( double x )
 {
 	char *buf;
 	int  i, dec, sig;
